% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmfunctions.R
\name{prepDFM}
\alias{prepDFM}
\title{Create DFM object}
\usage{
prepDFM(db = NULL, numbers = TRUE, punct = TRUE, stopwords = TRUE, symbols = TRUE, stem = TRUE, minchars = 3, textfield=NULL, idfield=NULL)
}
\arguments{
\item{db}{Dataframe with document text and ID variables}

\item{numbers}{Eliminate numbers from the DFM}

\item{punct}{Eliminate punction from the DFM}

\item{stopwords}{Eliminate stopwords from the DFM}

\item{symbols}{Eliminate symbols from the DFM}

\item{stem}{Stem tokens when creating the DFM}

\item{minchars}{Minimum number of characters for token to appear in DFM}

\item{textfield}{Name of text variable in dataframe}

\item{idfield}{Name of document id variable in dataframe}
}
\value{
Returns Quanteda DFM object
}
\description{
Create DFM object
}
