% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmfunctions.R
\name{prepLDA}
\alias{prepLDA}
\title{Create STM object from a DFM object}
\usage{
prepLDA(db = NULL, lt = 0.025, ut = 0.4)
}
\arguments{
\item{db}{Name of existing DFM object to convert to STM object}

\item{lt}{Lower threshold of document percentage (0 to 1 scale). Terms that occur in less than this percenage of documents will be dropped.}

\item{ut}{Upper threshold of document percentage (0 to 1 scale). Terms that occur in more than this percentage of documents will be dropped.}
}
\value{
Returns STM object
}
\description{
Create STM object from a DFM object
}
