% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmfunctions.R
\name{runSTM}
\alias{runSTM}
\title{Runs STM for given number of topics}
\usage{
runSTM(db = NULL, stmobj = NULL, k = NULL, ...)
}
\arguments{
\item{db}{Name to use to prefix saved files}

\item{stmobj}{STM object to run analysis on}

\item{k}{Number of topics to estimate}

\item{...}{Not used}
}
\description{
Runs STM for given number of topics
}
