% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmfunctions.R
\name{searchKbroad}
\alias{searchKbroad}
\title{Wrapper for STM's searchK function}
\usage{
searchKbroad(db = NULL, outfile = NULL, trange = NULL)
}
\arguments{
\item{db}{Name of STM object}

\item{outfile}{File location to save graph}

\item{trange}{Range of topic numbers to search}
}
\value{
Returns png file of searchK graph
}
\description{
Wrapper for STM's searchK function
}
